helper merge
